+++ tests/testthat/test-conformance.R
@@ -19,8 +19,6 @@ library(data.table)
 
 test_that("correct functioning of conformance", {
 
-  parallel(parallel = TRUE, logfile_dir = tempdir(), ncores = 2)
-
   utils_path <- system.file(
     "demo_data/utilities/",
     package = "DQAstats"