+++ tests/testthat/test-etl_and_differences_checks.R
@@ -19,8 +19,6 @@ library(data.table)
 
 test_that("correct functioning of etl_checks and difference_checks", {
 
-  parallel(parallel = TRUE, logfile_dir = tempdir(), ncores = 2)
-
   utils_path <- system.file(
     "demo_data/utilities/",
     package = "DQAstats"