+++ tests/testthat/test-helper_vars.R
@@ -19,8 +19,6 @@ library(data.table)
 
 test_that("correct functioning of helper vars", {
 
-  parallel(parallel = TRUE, logfile_dir = tempdir(), ncores = 2)
-
   source_system_name <- "exampleCSV_source"
   target_system_name <- "exampleCSV_target"